Add progress argument to all_transforms. Used to keep track of reduction progress across invocations of all_transforms.
Make TokenRemover aware of its progress. This significantly reduces the number of iterations required to reach a fixed point
since it no longer restarts with tokens that couldn't be removed in previous passes.

Significant improvement to overall reduction speed by moving computation of constants out of the innermost for-loop in the reducer.